feat: rename shifting field to local detuning
